Move Sanitizers to a submodule
